+++ tests/frontend/org/voltdb/regressionsuites/TestInitStartLocalClusterInProcess.java
@@ -63,7 +63,7 @@ public class TestInitStartLocalClusterInProcess extends JUnit4LocalClusterTest {
     String listener;
     Client client;
     String voltDbRootPath;
-    String voltDBRootParent;
+    String voltDBRootParentPath;
 
     @Before
     public void setUp() throws Exception {
@@ -89,7 +89,7 @@ public class TestInitStartLocalClusterInProcess extends JUnit4LocalClusterTest {
             voltDbRoot = new File(voltDbFilePrefix, builder.getPathToVoltRoot().getPath());
         }
         voltDbRootPath = voltDbRoot.getCanonicalPath();
-        voltDBRootParent = voltDbRoot.getParent();
+        voltDBRootParentPath = voltDbRoot.getParentFile().getCanonicalPath();
         listener = cluster.getListenerAddresses().get(0);
         client = ClientFactory.createClient();
         client.createConnection(listener);
@@ -118,11 +118,17 @@ public class TestInitStartLocalClusterInProcess extends JUnit4LocalClusterTest {
     }
 
     @Test
+    // Test get deployment
     public void testGetDeployment() throws Exception {
+        // get command is not supported in legacy cli as voltdbroot
+        // under the parent can't be determined deterministically
+        // using voltdbroot as the root of database directory
+        if (!cluster.isNewCli()) return;
+
         File deployment = File.createTempFile("get_deployment", ".xm");
 
         Configuration c1 = new VoltDB.Configuration(new String[]{"get", "deployment",
-            "getvoltdbroot", voltDBRootParent,
+            "getvoltdbroot", voltDBRootParentPath,
             "file", deployment.getAbsolutePath() + "l", "forceget"});
         ServerThread server = new ServerThread(c1);
 
@@ -138,12 +144,18 @@ public class TestInitStartLocalClusterInProcess extends JUnit4LocalClusterTest {
     }
 
     @Test
+    // Test get schema
     public void testGetSchema() throws Exception {
+        // get command is not supported in legacy cli as voltdbroot
+        // under the parent can't be determined deterministically
+        // using voltdbroot as the root of database directory
+        if (!cluster.isNewCli()) return;
+
         File schema = File.createTempFile("schema", ".sql");
         if (schema.exists()) schema.delete();
 
         Configuration c1 = new VoltDB.Configuration(new String[]{"get", "schema",
-            "getvoltdbroot", voltDBRootParent,
+            "getvoltdbroot", voltDBRootParentPath,
             "file", schema.getAbsolutePath()});
         ServerThread server = new ServerThread(c1);
